test(editor): colocate instruction envelope tests with EditorAgentNamespace
Move the "instruction envelope preservation (providerOptions)" suite from
apply-stored-overrides.test.ts into namespaces/agent.test.ts so it sits
next to the applyStoredOverrides implementation it covers. No behavior
changes to the tests themselves.

Addresses https://github.com/mastra-ai/mastra/pull/22493#discussion_r3875437183
